Removed all the slow calls to sum() in express_matrix_from_pixels_and_express()
One (inside a for loop) was replaced by a numpy array operation

The others were in a section of code that was Bad Idea anyway.
It had been an option in the IDL version of arctic. It hadn't made it to the C++ code.
Although it improved overall convergence at very low express, it did so at a cost of
local discontinuities in the gradients of charge trails (whenever the number of
evaluations increased by 1). It would also have prevented this python code from
reproducing the output of the C++ code when express>0.